backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-allauth>=0.57,<0.58
django-cors-headers>=4.3,<4.4
channels>=4.0,<4.1
channels-redis>=4.1,<4.2
daphne>=4.0,<4.1
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<2.2
Pillow>=10.0,<11.0
supabase>=2.0,<3.0
python-decouple>=3.8,<3.9
redis>=5.0,<5.1
